Move DataColumnType to extensions lib
