+++ lib/core/src/lib/datatable/data/object-datacolumn.model.ts
@@ -16,7 +16,8 @@
  */
 
 import { TemplateRef } from '@angular/core';
-import { CurrencyConfig, DataColumn, DataColumnType } from './data-column.model';
+import { DataColumnType } from '@alfresco/adf-extensions'
+import { CurrencyConfig, DataColumn } from './data-column.model';
 
 // Simple implementation of the DataColumn interface.
 export class ObjectDataColumn<T = unknown> implements DataColumn<T> {
