+++ lib/extensions/src/lib/config/document-list.extensions.ts
@@ -15,8 +15,22 @@
  * limitations under the License.
  */
 
-import { DataColumnType } from '@alfresco/adf-core';
 import { ExtensionElement } from './extension-element';
+
+export interface DataColumnTypes {
+    text: string;
+    image: string;
+    date: string;
+    json: string;
+    icon: string;
+    fileSize: string;
+    location: string;
+    // eslint-disable-next-line id-blacklist
+    boolean: string;
+    amount: string;
+}
+
+export type DataColumnType = keyof DataColumnTypes;
 
 export interface DocumentListPresetRef extends ExtensionElement {
   key: string;
